+++ src/types.ts
@@ -1,0 +1,8 @@
+interface Organization {
+    id: string, 
+    ownerId: string,
+    name: string,
+    description: string
+}
+
+export type { Organization };
